fix: honor numeric display_format in describe/diff/netlist (hex constants)
describe/diff/netlist printed numeric constants in decimal where the renderer
shows LabVIEW's radix (MD5 init 0x67452301 showed as 1732584193). The display
format lives on ConstantNode (parser-populated) but the slim Constant
projection built by get_constants (queries.py) dropped it — so display commands
that ride on that projection lost it, while render (which walks the nodes) kept it.

- Move the renderer's radix/precision formatter into a neutral lvkit/num_format.py
  (graph can't import render), shared by render and the graph/netlist text path.
- Add display_format to the Constant model + thread it through get_constants.
- _const_value_str honors it, so describe/diff/netlist match the renderer.

Codegen is unaffected — it emits constants via a separate path, still valid
Python decimal. All tests green. See task #73 for the underlying access-pattern
standardization.
